Update v0.1.5
Okay, the game just became a little harder.

Now the zombies can now detect the play and take health. The game also
end when you lose health.

The start/ and gameover functions need to be cleaned up. So does the
zombie detection a little bit.

This is a pre release. I would not say it it ready to move to 0.1.6 More
on the way.